Migration: refactor _checkExpressionNotNull.
Previously this was based on creating a dummy type pointing to `never`
and then trying to assign to it.  Since we are trying to get rid of
explicit references to `never` outside of the graph logic, that isn't
going to work anymore.  Refactor to just create the correct edge
directly using NullabilityGraph.makeNonNullable.

Change-Id: I18683bfe5e02924d2ccfcb0459047568734f975d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/125801
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>