[nnbd_migration] Only restyle code links (which I also bolded)
This leaves default link styling for links that aren't click-thru nav,
so inside the pop-ups as well as the links to other files at the top of
the page. Default styling as in blue and underlined.

I also made the code links bold when hovered, because the participant
didn't try clicking on them. Hopefully this grabs more attention.

Change-Id: Ib35c2353e54faa19ba2fa8e0a613ecabd3298463
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/126206
Auto-Submit: Mike Fairhurst <mfairhurst@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>