[CFE] [Strong test] Check non-serialized expectations first
This gives a "MismatchExpectations" error instead of
"MismatchExpectationsSerialized" if the serialization doesn't matter in
getting to the mismatch.
The serialized version is skipped when updating expectations to not
makeing the serialized version the "right" verison (which could get
you into a position where updating expectations and rerunning would
tell you the (normal) expectations needed updating).

Change-Id: I99b3af5fa823010d4abbf0837faf00e17b303c0c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/126540
Reviewed-by: Dmitry Stefantsov <dmitryas@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>