[ package:vm_service ] Fix issue where an exception could be thrown if remote service protocl version was older than that supported by package:vm_service
When connecting to a VM service which supports an older version of the
protocol than that supported by package:vm_service, there was the
possibility for a crash if the package tried to create a List for a
response property which is not supported by the remote VM service. The
fix here is to check for null from createServiceObject when building a
List from a response and produce an empty list for that property.

Issue originally filed here: https://github.com/dart-lang/coverage/issues/278

Change-Id: Ibe8593fa8e4a55d05dd821bbdf363a67c15d5483
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/127065
Reviewed-by: Devon Carew <devoncarew@google.com>
Commit-Queue: Ben Konyi <bkonyi@google.com>