Create an abstract layer to customize querying element types.
This will be needed for the NNBD migration engine.

Currently only used in checker.dart.  Later CLs will convert the
resolver to use this abstraction layer as well.

The runtime overhead of this abstraction should be negligible, since
under normal analyzer operation, there will only be a single const
instance of this class, and no instances of any subclasses.  So the VM
should be able to trivially inline all uses of the abstraction layer.

Change-Id: I0132b76238606e6e6b789169a62806d09f10f71d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/127490
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>