[infra] Add failing test to test CI systems and approvals workflow
The failing test lib_2/collection/failing_list_test.dart is added,
which imports dart:io.  So it will be a compile-time failure on
some dart2js platforms, and a run-time failure on dart:io.

This commit is just to test the UI and backend for test failure
approvals on tryjobs and on CI results. It will be reverted after
the CI builders have run on it.

Change-Id: Icc3d11e72511cb0dbd9a62cb4fa80e754c203c8c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/127883
Reviewed-by: Jonas Termansen <sortie@google.com>