Widen the analyzer's pubspec constraint on _fe_analyzer_shared
Note that this means we now need to be careful to bump the version of
package:_fe_analyzer_shared when making breaking changes to its API.
Hopefully when making a breaking change we can remember to bump
package:_fe_analyzer_shared to a "-dev" version.  However, as a safety
check, I've added a note to my process for publishing
analyzer/_fe_analyzer_shared to pub, to double-check for breaking
changes and update version numbers appropriately.

This should help address #39711.

Bug: https://github.com/dart-lang/sdk/issues/39711
Change-Id: Icd942bc4e7dc619d44f00369ed447e0cffb8d712
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/127682
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>