(dart2js): escape names when creating tearoff closures
The CFE recently changed the name of the top-level deferred load library method
from `__loadLibrary` to `_#loadLibrary` (see #39682)

We were properly escaping this name when generating the method itself, but we
didn't properly escape the name when creating tearoff closures derived from it.

This is why the issue was not detected earlier.

Change-Id: Ic1487a10f521a86a3466608b1825a2b632a191be
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/128260
Reviewed-by: Stephen Adams <sra@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>