Create an interface for the migration engine to communicate with the resolver.
The new interface (MigrationResolutionHooks) extends the
ElementTypeProvider interface introduced in previous CLs.

This CL also contains plumbing to allow the resolver to be created
using the new interface.  In a follow-up CL I will add code to the
migration engine that implements the new interface and passes it into
the resolver.

Change-Id: I717d71e3c6dc65f136d89ca1c2a10b3f89f2f356
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/128483
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>