[dartdevc] Add an nnbd runtime mode weak/strong option to ddb script
Cleaned up the help message printing and sorted the options alphabetically
since the list is getting longer.

Makes it easier to run a file in the various modes. Will automatically pass
the experiment flag to dartdevc if you run with weak or strong.

Added a verbose mode to print the commands being run with all the args and
environment.

Change-Id: I515d185d940b0a13ac74b8b58ae1510a5b203ae7
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/128775
Reviewed-by: Mark Zhou <markzipan@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>