[nnbd_migration] fix crash by composing non-null/type promotion
There were cases where types that are promoted with current type
promotion were not being promoted with improved type promotion due to
null-checks:

```dart
void f(x) {
  if (x == null) {
    return;
  }

  if (x is ...) ...
}
```

This was causing a bad state when trying to substitute type parameters
on methods if promotion usually succeeded.

Fix the crash by allowing migration to promote a non-null type to a new
type, and make that new type an intersection of the new type and
non-null.

```dart
  if (x == null) {
  } else if (x is List<int/*1*/>/*2*/) { // promote to List<int/*1*/>!
```

Change-Id: Icf9f74eb2b3eef887a29d4bfab7556ac126f49e4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/129540
Commit-Queue: Mike Fairhurst <mfairhurst@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>